long int

+++ customer.h
@@ -11,7 +11,7 @@ class Customer
 public:
     string firstName;
     string FamilyName;
-    long int ElectricityAccountId;
+    long long int ElectricityAccountId;
     int BankAccount;
     int familyMembersNumber;
     int *Ages;
@@ -27,11 +27,11 @@ public:
     int id;
 
     Customer() {}
-    Customer(string, string, int, int, int *A, string, string, string, long int);
+    Customer(string, string, int, int, int *A, string, string, string,long long int);
     void setInfo(string, string, int, int, int *A, string, string, string);
     vector<string> getIDs(string, string, string);
     string getConcatenatedIDs(string, string, string);
-    long int generateCustomerID(string, string, string, long int);
+    long int generateCustomerID(string, string, string, long long int);
     void settotalInjection(int);
     long int getCustomerId();
     void getOneMonthBillCustomer(int, int);
